Add new `render_screen` function which you MUST call
This is how nene drawing works currently:

There is two different render important targets (if you don't setup any
new one): the "window", and the "screen".

Screen is made for game rendering, which have a specific resolution,
while the "window" is "no render target", which is the window directly.

`pool_events` already sets the screen as the render target, and
`render_screen` resets the render target to window and renders the
screen on it.

`render present` presents the final result to the player.
---
The goal is that any game-related drawing should be done before
`render_screen` while any GUI-drawing should be done after it, of course
you are free to do anything you want.

note that what really matters is the resolution, the "screen" may have a
different resolution compared to window's resolution, for example, in
Celeste, the game is made with pixel art, while the GUI is made in
high-resolution.