`Nene.Core` rework
The major change is that `Nene.Core` really acts more like a "Core",
that is, it doesn't depends on any other Nene module (except `Nene.Math`
and `Nene.Color`) and only holds the minimum information it needs to
work.

The removed functions will be moved to others modules in future commits.

Signed-off-by: André Luiz Alvares <andrluiz.alvares.alvares84@gmail.com>