using nldoc and big refactor about modularization
`Nene` is not a namespace anymore, but it replaces `Nene.Core`.

Also, every module now declares a global, without any namespace.

The reason is that, nelua will get modules implemented eventually, so it
doesn't makes much sense to add namespacing for the whole Nene since it
would be removed.

While not all modules are actually refactored, the Pong example works.

The documentation also need a review, so the docs may be wrong now.