add "zero" and "copy" on all modules.
This adds a language-independent zero initialization and data copying, on some languages this might be useless, but can help others (like Carp in this case).

Note that's just a shallow copy (honestly a dereference from the pointer), not a deep copy, so calling copy from a resource like Font will not load another Font from disk, it'll just copy the handle.